Statespcemodel and filter
